+++ src/editor/SceneHierarchyPanel.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "EditorPanel.h"
+
+class SceneHierarchyPanel : public EditorPanel
+{
+public:
+    SceneHierarchyPanel();
+    void OnImGuiRender(float dt) override;
+};